All recognized serology are valid. Applied black.
